+++ demos/zmq_rr/actors/sample_processor.py
@@ -2,7 +2,7 @@ from improv.actor import Actor, AsyncActor
 import numpy as np
 import logging
 
-from demos.basic.actors.zmqActor import ZmqPSActor, ZmqRRActor
+from demos.sample_actors.zmqActor import ZmqPSActor, ZmqRRActor, ZmqActor
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -32,7 +32,8 @@ class Processor(Actor):
         self.frame = None
         self.avg_list = []
         self.frame_num = 1
-        self.subscribe = ZmqRRActor("processor", self.store_loc)
+        #self.subscribe = ZmqRRActor("processor", self.store_loc)
+        self.subscribe = ZmqActor("processor", self.store_loc, pub_sub=False, rep_req=True)
         logger.info("Completed setup for Processor")
 
     def stop(self):
