+++ pyproject.toml
@@ -44,6 +44,10 @@ exclude = ["test", "pytest", "env", "demos", "figures"]
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 filterwarnings = [ ]
+log_cli = true
+log_cli_level = "INFO"
+log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
+log_cli_date_format = "%Y-%m-%d %H:%M:%S"
 
 [tool.flake8]
 # Recommend matching the black line length (default 88),
@@ -69,3 +73,5 @@ exclude = '''
   | build
 )/
 '''
+
+
